+++ plugins_init.h
@@ -34,9 +34,11 @@
 #endif
 
 #if VFD_ENABLE
-    extern void vfd_init (void);
-    vfd_init();
-#if DUAL_SPINDLE
+#if VFD_ENABLE == 1 || VFD_ENABLE == 2 || VFD_ENABLE == -1
+    extern void vfd_huanyang_init (void);
+    vfd_huanyang_init();
+#endif
+#if N_SPINDLE > 1
     extern void spindle_select_init(void);
     spindle_select_init();
 #endif